Story content

+++ back-end-api/fluencybox/S3AssetManager.py
@@ -4,9 +4,9 @@ import os, io
 import boto3
 from fluencybox import app
 from botocore.exceptions import NoCredentialsError
-from fluencybox.config import S3_BUCKET, S3_KEY, S3_SECRET
+from fluencybox.config import S3_BUCKET, S3_KEY, S3_SECRET, SQS_QUEUE_URL, S3_REGION
 import mimetypes
-
+import json
 def get_resource():
     if S3_KEY and S3_SECRET:
         return boto3.resource(
@@ -33,22 +33,27 @@ def get_client():
                       aws_secret_access_key=S3_SECRET)
     return s3
 
-def save_s3_object(object_dir, body, content_type):
+def save_s3_object(object_dir, body, content_type, is_public = False):
     try:
         resp_dict = {}
+        object_url = ''
         my_bucket = get_bucket()
-        s3 = get_resource()
-        #set directory, object and content type of the object
-        my_bucket.Object(object_dir).put(Body = body, ContentType = content_type)
-        #set access permissions
-        object_acl = s3.ObjectAcl(app.config.get('S3_BUCKET'), object_dir)
-        response = object_acl.put(ACL='public-read')
-        #get full URL to the object
-        object_url = app.config.get('S3_URL') + object_dir
+
+        if is_public:
+            upload_response = my_bucket.Object(object_dir).put(Body = body, ContentType = content_type, ACL='public-read') #set directory, object, content type & public access of the object
+            if upload_response['ResponseMetadata']['HTTPStatusCode'] == 200:
+                object_url = app.config.get('S3_URL') + object_dir #get full URL to the object
+        else:
+            upload_response = my_bucket.Object(object_dir).put(Body = body, ContentType = content_type) #set directory, object and content type of the object
         
-        resp_dict['status'] = 'success'
-        resp_dict['object_url'] = object_url
-        return resp_dict
+        if upload_response['ResponseMetadata']['HTTPStatusCode'] == 200:
+            resp_dict['status'] = 'success'
+            resp_dict['object_url'] = object_url
+            return resp_dict
+        else:
+            resp_dict['status'] = 'fail'
+            resp_dict['message'] = upload_response['ResponseMetadata']['HTTPStatusCode']
+            return resp_dict
 
     except Exception as e:
         resp_dict['status'] = 'fail'
@@ -65,12 +70,16 @@ def save_avatar(my_avatar):
         mime_type = mimetypes.types_map[obj_ext]
 
         object_dir = app.config.get('S3_AVATAR_DIR') + '/' + obj_fn
-        object_url = save_s3_object(object_dir, my_avatar, mime_type)
+        object_url = save_s3_object(object_dir, my_avatar, mime_type, True)
 
         if object_url['status'] == 'success':
             resp_dict['status'] = 'success'
             avatar_url = object_url['object_url']
             resp_dict['object_url'] = avatar_url
+        else:
+            resp_dict['status'] = 'fail'
+            resp_dict['message'] = object_url['message']
+            return resp_dict
 
         return resp_dict
 
@@ -79,24 +88,22 @@ def save_avatar(my_avatar):
         resp_dict['message'] = str(e)
         return resp_dict
 
-def save_story_object(my_object):
+def save_story_object(my_object, object_filename, is_public = False):
     try:
         resp_dict = {}
-        
-        _, obj_ext = os.path.splitext(my_object.filename)
+        _, obj_ext = os.path.splitext(object_filename) 
         mime_type = mimetypes.types_map[obj_ext]
-
-        object_dir = app.config.get('S3_CONTENT_DIR') + '/' + my_object.filename
-        object_url = save_s3_object(object_dir, my_object, mime_type)
-
+        object_url = save_s3_object(object_filename, my_object, mime_type, is_public)
         if object_url['status'] == 'success':
             resp_dict['status'] = 'success'
             story_object_url = object_url['object_url']
             resp_dict['object_url'] = story_object_url
-
-        resp_dict['status'] = 'success'
-        resp_dict['object_url'] = object_url
-        return resp_dict
+            return resp_dict
+        else:
+            resp_dict['status'] = 'fail'
+            resp_dict['message'] = object_url['message']
+            return resp_dict
+        
     except Exception as e:
         resp_dict['status'] = 'fail'
         resp_dict['message'] = str(e)
@@ -108,7 +115,9 @@ def delete_avatar(my_avatar):
         my_bucket = get_bucket()
         file_name = my_avatar.split('/')[-1]
         file_name = app.config.get('S3_AVATAR_DIR') + '/' + file_name
-        my_bucket.Object(file_name).delete()
+        default_image = app.config.get('S3_AVATAR_DIR') + '/' + app.config.get('DEFAULT_IMAGE')
+        if file_name != default_image:
+            my_bucket.Object(file_name).delete()
         
         resp_dict['status'] = 'success'
         resp_dict['message'] = 'successfully deleted'
@@ -124,11 +133,68 @@ def delete_story_object(my_object_key):
         resp_dict = {}
         my_bucket = get_bucket()
         file_name = my_object_key.split('/')[-1]
-        file_name = app.config.get('S3_CONTENT_DIR') + '/' + file_name
         my_bucket.Object(file_name).delete()
         
         resp_dict['status'] = 'success'
         resp_dict['message'] = 'successfully deleted'
+        return resp_dict
+
+    except Exception as e:
+        resp_dict['status'] = 'fail'
+        resp_dict['message'] = str(e)
+        return resp_dict
+
+def generate_presigned_url(bucket, key):
+    resp_dict = {}
+    s3_client = get_client()
+    try:
+        expiry = int(app.config['SIGNED_URL_EXPIRY'])
+        if expiry < 1:
+            expiry = 300
+    except Exception as e:
+        expiry = 300
+
+    url = s3_client.generate_presigned_url(ClientMethod = 'get_object', Params = {'Bucket': bucket, 'Key': key}, ExpiresIn = expiry)
+    return url
+
+def generate_public_url(object_type, object_name):
+    if object_type == 'story_image':
+        object_dir = app.config.get('STORY_IMAGES_DIR')
+    elif object_type == 'speaker_image':
+        object_dir = app.config.get('SPEAKER_IMAGES_DIR')
+    elif object_type == 'speaker_audio':
+        object_dir = app.config.get('SPEAKER_AUDIO_DIR')
+    elif object_type == 'master_audio':
+        object_dir = app.config.get('MASTER_RESPONSE_AUDIO_DIR')
+    elif object_type == 'user_audio':
+        object_dir = app.config.get('USER_RESPONSE_AUDIO_DIR')
+    elif object_type == 'report_image':
+        object_dir = app.config.get('REPORT_IMAGES_DIR')
+
+    public_url = app.config.get('S3_URL') + object_dir + '/' + object_name
+    return public_url
+
+def get_sqs_client():
+    sqs = boto3.client('sqs', aws_access_key_id=S3_KEY, aws_secret_access_key=S3_SECRET, region_name = S3_REGION)
+    return sqs
+
+
+def trigger_sqs(payload):
+    try:
+        resp_dict = {}
+       # Create SQS client
+        sqs = get_sqs_client()
+        
+        # Send message to SQS queue
+        response = sqs.send_message(
+            QueueUrl=SQS_QUEUE_URL, 
+            MessageBody=(payload)
+            )
+
+        print(response['MessageId'])
+
+        resp_dict['status'] = 'success'
+        resp_dict['message'] = 'sqs triggered'
         return resp_dict
 
     except Exception as e:
